+++ crates/jazz-rn/CHANGELOG.md
@@ -1,5 +1,9 @@
 # cojson-core-rn
 
+## 2.0.0-alpha.24
+
+## 2.0.0-alpha.23
+
 ## 2.0.0-alpha.22
 
 ### Patch Changes